libs and hack to use em
I want the chance that people can play with this, even in it's early
broken state so I've thrown in the build libs and made a seperate asd
which doesnt try to use c2ffi. The reason for that is that cffi-c2ffi
doesnt handle unamed fields in some datatypes correctly and so I've
hacked it into a working state.